requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.8
dj-database-url==2.3.0
dj-static==0.0.6
Django==5.1.5
django-bootstrap-form==3.4
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
pillow==11.1.0
python-decouple==3.8
setuptools==75.8.0
sqlparse==0.5.3
static3==0.7.0
typing_extensions==4.12.2
uvicorn==0.34.0
whitenoise==6.8.2
